requirements.txt
flask
pandas
numpy
tensorflow
scikit-learn
python-dotenv
requests
gunicorn